unified-proxy/requirements.txt
# Unified Proxy Python Dependencies
#
# Core proxy framework (also provides mitmproxy Python API)
mitmproxy~=11.0

# Web framework for internal API (container registration, metrics)
flask~=3.0

# HTTP client for OAuth token refresh and external API calls
httpx~=0.27

# Async HTTP server + client for API gateway endpoints
aiohttp~=3.9

# YAML parsing for configuration files
pyyaml~=6.0

# Prometheus metrics collection and exposition
prometheus-client~=0.20

# Zstandard decompression support for aiohttp (required by Codex CLI)
backports.zstd
